+++ examples/perplexity-chat/src/main.rs
@@ -1,5 +1,5 @@
 use perplexity::{
-    client::{Client, CreateChatCompletion, Model},
+    client::{Client, CreateChatCompletion, Message, Model, Role},
     config::Config,
 };
 
@@ -11,7 +11,11 @@ async fn main() {
     let config = Config::new(api_key);
     let client = Client::new(config).unwrap();
 
-    let message = CreateChatCompletion::new(Model::Llama31SonarLargeOnline);
+    let messages: Vec<Message> = vec![Message {
+        role: Role::User,
+        content: "Find me the best pad thai restaurant in Bangkok".to_string(),
+    }];
+    let message = CreateChatCompletion::new(Model::Llama31SonarLargeOnline, messages);
     let result = client.create_completion(message).await.unwrap();
     println!("{:?}", result);
 }